pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "torque-tunnel"
version = "0.1.0"
description = "MCP tool for executing remote commands via Torque tunneled infrastructure"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
torque-tunnel = "torque_tunnel.mcp_tool:main"

[tool.setuptools.packages.find]
where = ["src"]
